no 3d gif without catalog

+++ online_skyloc/threeDvolume.py
@@ -345,13 +345,14 @@ class VolumeReconstruction(mixture):
         
     def make_gif(self):
         files = [f for f in self.gif_folder.glob('3d_'+self.name + '*' + '.png')]
-        path_files = [str(f) for f in files]
-        path_files.sort(key = natural_keys)
-        images = []
-        for file in path_files:
-            images.append(imageio.imread(file))
-        imageio.mimsave(Path(self.gif_folder, '3d_'+self.name + '.gif'), images, fps = 1)
-        [f.unlink() for f in files]
+        if len(files) > 0:
+            path_files = [str(f) for f in files]
+            path_files.sort(key = natural_keys)
+            images = []
+            for file in path_files:
+                images.append(imageio.imread(file))
+            imageio.mimsave(Path(self.gif_folder, '3d_'+self.name + '.gif'), images, fps = 1)
+            [f.unlink() for f in files]
 
         files = [f for f in self.gif_folder.glob(self.name + '*' + '.png')]
         path_files = [str(f) for f in files]
